requirements.txt
allure-behave==2.13.5
allure-python-commons==2.13.5
Appium-Python-Client==2.7.1
attrs==23.2.0
behave==1.2.6
certifi==2024.2.2
cffi==1.16.0
cryptography==42.0.7
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.0
parse==1.20.1
parse-type==0.6.2
pluggy==1.5.0
pycparser==2.22
pyOpenSSL==24.1.0
PySocks==1.7.1
pytest==8.2.0
selenium==4.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==1.26.18
urllib3-secure-extra==0.1.0
wsproto==1.2.0
